fix(dts): 修复部分组件事件类型丢失问题
