+++ src/packages/__VUE/skeleton/index.taro.vue
@@ -17,12 +17,106 @@
 
 <script lang="ts">
 import { createComponent } from '@/packages/utils/create';
-import { component } from './common';
 import Avatar from '../avatar/index.taro.vue';
+import { AvatarShape } from '../avatar/types';
+import { CSSProperties, PropType, computed, toRefs } from 'vue';
 const { create } = createComponent('skeleton');
-export default create(
-  component({
+export default create({
+  components: {
     [Avatar.name]: Avatar
-  })
-);
+  },
+  props: {
+    //每行宽度
+    width: {
+      type: String,
+      default: '100px'
+    },
+    //每行高度
+    height: {
+      type: String,
+      default: '15px'
+    },
+    //是否显示动画
+    animated: {
+      type: Boolean,
+      default: false
+    },
+    //头像
+    avatar: {
+      type: Boolean,
+      default: false
+    },
+    //头像形状：正方形/圆形
+    avatarShape: {
+      type: String as PropType<AvatarShape>,
+      default: 'round'
+    },
+    //头像大小
+    avatarSize: {
+      type: String,
+      default: '50px'
+    },
+    //是否显示骨架屏
+    loading: {
+      type: Boolean,
+      default: true
+    },
+    //标题/段落 圆角风格
+    round: {
+      type: Boolean,
+      default: false
+    },
+
+    //显示段落行数
+    row: {
+      type: String,
+      default: '1'
+    },
+
+    //是否显示段落标题
+    title: {
+      type: Boolean,
+      default: true
+    }
+  },
+
+  setup(props: any) {
+    const { avatarShape, round, avatarSize } = toRefs(props);
+
+    const avatarClass = computed(() => {
+      const prefixCls = 'avatarClass';
+      return {
+        [prefixCls]: true,
+        [`${prefixCls}--${avatarShape.value}`]: avatarShape.value
+      };
+    });
+
+    const getBlockClass = (prefixCls: string) => {
+      return {
+        [prefixCls]: true,
+        [`${prefixCls}--round`]: round.value
+      };
+    };
+
+    const getStyle = (): CSSProperties => {
+      if (avatarSize.value) {
+        return {
+          width: avatarSize.value,
+          height: avatarSize.value
+        };
+      }
+      return {
+        width: '50px',
+        height: '50px'
+      };
+    };
+
+    return {
+      avatarShape,
+      avatarClass,
+      getBlockClass,
+      getStyle
+    };
+  }
+});
 </script>
